requirements.txt
alembic==1.4.3
astroid==2.4.2
attrs==20.2.0
cliff==3.4.0
cmaes==0.6.1
cmd2==1.3.11
colorama==0.4.4
colorlog==4.4.0
cupy-cuda110==8.0.0
fastrlock==0.5
isort==5.6.4
joblib==0.17.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.2
optuna==2.2.0
packaging==20.4
pbr==5.5.0
prettytable==0.7.2
pylint==2.6.0
pyparsing==2.4.7
pyperclip==1.8.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
scipy==1.5.3
six==1.15.0
SQLAlchemy==1.3.20
stevedore==3.2.2
toml==0.10.1
tqdm==4.50.2
wcwidth==0.2.5
wrapt==1.12.1
